version running on ROS2 - eloquent: `ros2 launch realsense2_camera rs.launch.py`
missing:
- runtime configurable parameters.
- T265
- realsense2_description
- node namespace in topic names.

+++ realsense2_camera/include/constants.h
@@ -48,9 +48,9 @@ namespace realsense2_camera
 
     const int IMAGE_WIDTH     = 640;
     const int IMAGE_HEIGHT    = 480;
-    const int IMAGE_FPS       = 30;
+    const double IMAGE_FPS    = 30;
 
-    const int IMU_FPS         = 0;
+    const double IMU_FPS      = 0;
 
 
     const bool ENABLE_DEPTH   = true;
